Delete function on admin site at watch table
