+++ watches/app/Http/Controllers/Admin/AdminController.php
@@ -95,18 +95,18 @@ class AdminController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $request)
+    public function destroy (Request $request)
     {
-        // validated -- make sure id is passed in request
-        $valid = $request->validate([
-            'id' => 'required|integer'
-        ]);
+        // // validated -- make sure id is passed in request
+        // $valid = $request->validate([
+        //     'customer_id' => 'required|integer'
+        // ]);
 
-        // Try to delete the post and send the user back to the posts 
-        // index view with a flash message
-        if( Customers::find($valid['id'])->delete() ) {
-            return back()->with('success', 'Post has been deleted!');
-        }
-        return back()->with('error', 'There was a problem deleting that post');
+        // // Try to delete the post and send the user back to the posts 
+        // // index view with a flash message
+        // if( User::find($valid['customer_id'])->delete() ) {
+        //     return back()->with('success', 'Post has been deleted!');
+        // }
+        // return back()->with('error', 'There was a problem deleting that post');
     }
 }
