+++ watches/app/Http/Controllers/Admin/CustomerController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
+use App\Admin; 
+use App\Customer; 
+
 class CustomerController extends Controller
 {
     /**
@@ -78,8 +81,18 @@ class CustomerController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request)
     {
-        //
+        // validated -- make sure id is passed in request
+        $valid = $request->validate([
+            'customer_id' => 'required|integer'
+        ]);
+
+        // Try to delete the post and send the user back to the posts 
+        // index view with a flash message
+        if( Customer::find($valid['customer_id'])->delete() ) {
+            return back()->with('success', 'Post has been deleted!');
+        }
+        return back()->with('error', 'There was a problem deleting that post');
     }
 }
