+++ watches/resources/views/admin/watches_table.blade.php
@@ -33,7 +33,19 @@
 	      <td>$ {{ $watch->cost }}</td>
 	      <td>{{ $watch->material }}</td>
 	      <td> <p><a href="/admin/edit/{{ $watch->watch_id }}/edit_watch" class="btn btn-primary">Edit</a></p>
-	      <td><button type="button" class="btn btn-danger">Delete</button></td>
+	      <!-- <td><button type="button" class="btn btn-danger">Delete</button></td> -->
+
+
+	      <td>
+		  	<form class="delete" 
+                            onSubmit="return confirm('Do you really want to delete this post?')"
+                             action="/admin/watches_table" method="post">
+                            @csrf
+                            @method('DELETE')
+                            <input type="hidden" name="id" value="{{ $watch->id }}"/>
+                            <button type="submit" class="btn btn-danger">delete</button>
+            </form>
+         </td>
